Remove install script
Deno install copies the url as-is, there's no resolution, so installing a versionless url means always running the newest version, which isn't necessarily a good idea. If I'm going to have to include the version, might as well have it front and centre in the README instead of tucked away in an install script.